feat: add TIME_HANDLING.md documentation and improve type handling across scripts
Introduced a new documentation file, TIME_HANDLING.md, detailing conventions for time handling and primitive types. Conducted a comprehensive audit of type usage, fixing critical issues such as integer division in MTD calculations and float promotion in fingerprint hashing. Updated various scripts to utilize Math.MaxInt and Math.ClampInt for better type safety, ensuring consistent handling of integer values. Enhanced README and I18N documentation to include references to the new time handling guidelines.

+++ scripts/Game/RDF/Radar/Demo/RDF_RadarPromoReel.c
@@ -123,7 +123,7 @@ class RDF_RadarPromoReel
     protected static const float PROMO_MIN_RANGE_M = 50.0;
     protected static const float SHELL_IMPACT_HOLD_S = 3.5;
     protected static const float SHELL_TOF_FALLBACK_S = 18.5;
-    protected static const float TICK_MS = 33.0;
+    protected static const int TICK_MS = 33;
 
     protected static ref RDF_RadarPromoReel s_Instance;
     protected static bool s_TickRegistered;
